Remove debugging code and moved hook to seperate file
