+++ nerds-dating/src/components/ChatDisplay.jsx
@@ -31,7 +31,7 @@ const ChatDisplay = () => {
     return () => {
       socket.disconnect();
     };
-  });
+  },[]);
 
   return (
     <div className="flex flex-row px-8">
